+++ packages/typespec-ts/src/modular/buildCodeModel.ts
@@ -27,7 +27,6 @@ import {
   isStringType,
   getPropertyType,
   isNumericType,
-  getFormat,
   getMinItems,
   getMaxItems,
   listServices,
@@ -37,7 +36,8 @@ import {
   getProjectedName,
   isNullType,
   getEncode,
-  isTemplateDeclarationOrInstance
+  isTemplateDeclarationOrInstance,
+  getFormat
 } from "@typespec/compiler";
 import {
   getAuthentication,
@@ -52,7 +52,6 @@ import {
   HttpServer,
   isStatusCode,
   HttpOperation,
-  isHeader,
   getHttpOperation
 } from "@typespec/http";
 import { getAddedOnVersions } from "@typespec/versioning";
@@ -66,8 +65,9 @@ import {
   getClientNamespaceString,
   getSdkUnion,
   getAllModels,
-  SdkSimpleType,
-  getSdkSimpleType
+  SdkBuiltInType,
+  getSdkBuiltInType,
+  getClientType
 } from "@azure-tools/typespec-client-generator-core";
 import { getResourceOperation } from "@typespec/rest";
 import {
@@ -94,7 +94,8 @@ import {
 import {
   getOperationGroupName,
   getOperationName,
-  isIgnoredHeaderParam
+  isIgnoredHeaderParam,
+  isLongRunningOperation
 } from "../utils/operationUtil.js";
 import { SdkContext } from "../utils/interfaces.js";
 import { Project } from "ts-morph";
@@ -155,7 +156,8 @@ function isSimpleType(
       getMaxValue,
       getMinLength,
       getMaxLength,
-      getPattern
+      getPattern,
+      getEncode
     ];
     for (const func of funcs) {
       if (func(program, type)) {
@@ -170,47 +172,56 @@ function getDocStr(program: Program, target: Type): string {
   return getDoc(program, target) ?? "";
 }
 
-function isLro(_program: Program, operation: Operation): boolean {
-  for (const decorator of operation.decorators) {
-    if (decorator.decorator.name === "$pollingOperation") {
-      return true;
-    }
+function isDiscriminator(
+  context: SdkContext,
+  type: Model,
+  propertyName: string
+): boolean {
+  const discriminator = getDiscriminator(context.program, type);
+  if (discriminator && discriminator.propertyName === propertyName) {
+    return true;
   }
   return false;
 }
 
-function handleDiscriminator(context: SdkContext, type: Model, model: any) {
+function handleDiscriminator(context: SdkContext, type: Model) {
   const discriminator = getDiscriminator(context.program, type);
   if (discriminator) {
-    let discriminatorProperty;
+    const discriminatorValues: string[] = [];
     for (const childModel of type.derivedModels) {
       const modelType = getType(context, childModel);
       for (const property of modelType.properties) {
         if (property.restApiName === discriminator.propertyName) {
           modelType.discriminatorValue = property.type.value;
-          property.isDiscriminator = true;
-          model.discriminatedSubtypes[property.type.value] = modelType;
-          discriminatorProperty = property;
+          discriminatorValues.push(modelType.discriminatorValue);
         }
       }
     }
     // it is not included in properties of typespec but needed by python codegen
-    if (discriminatorProperty) {
-      const discriminatorType = { ...discriminatorProperty.type };
-      discriminatorType.value = null;
-      const propertyCopy = {
-        ...discriminatorProperty,
+    if (discriminatorValues.length > 0) {
+      const discriminatorInfo = {
+        description: `the discriminator possible values ${discriminatorValues.join(
+          ", "
+        )}`,
         isPolymorphic: true,
-        type: discriminatorType
+        isDiscriminator: true
       };
-      propertyCopy.description = "";
-      model.properties.push(propertyCopy);
+      return discriminatorInfo;
     }
   }
+  return undefined;
+}
+
+function isSchemaProperty(program: Program, property: ModelProperty): boolean {
+  const headerInfo = getHeaderFieldName(program, property);
+  const queryInfo = getQueryParamName(program, property);
+  const pathInfo = getPathParamName(program, property);
+  const statusCodeinfo = isStatusCode(program, property);
+  return !(headerInfo || queryInfo || pathInfo || statusCodeinfo);
 }
 
 function getEffectiveSchemaType(program: Program, type: Model | Union): Model {
-  function isSchemaProperty(property: ModelProperty) {
+  function isSchemaProperty(property: ModelProperty): boolean {
     const headerInfo = getHeaderFieldName(program, property);
     const queryInfo = getQueryParamName(program, property);
     const pathInfo = getPathParamName(program, property);
@@ -255,20 +266,20 @@ function processModelProperties(
 ) {
   // need to do properties after insertion to avoid infinite recursion
   for (const property of model.properties.values()) {
-    if (
-      isStatusCode(context.program, property) ||
-      isNeverType(property.type) ||
-      isHeader(context.program, property)
-    ) {
+    if (!isSchemaProperty(context.program, property)) {
       continue;
     }
     if (newValue.properties === undefined || newValue.properties === null) {
       newValue.properties = [];
     }
-    newValue.properties.push(emitProperty(context, property));
+    let newProperty = emitProperty(context, property);
+    if (isDiscriminator(context, model, property.name)) {
+      newProperty = { ...newProperty, ...handleDiscriminator(context, model) };
+    }
+    newValue.properties.push(newProperty);
   }
   // need to do discriminator outside `emitModel` to avoid infinite recursion
-  handleDiscriminator(context, model, newValue);
+  // handleDiscriminator(context, model, newValue);
 }
 
 function getType(
@@ -277,8 +288,7 @@ function getType(
   options: { disableEffectiveModel?: boolean } = {}
 ): any {
   // don't cache simple type(string, int, etc) since decorators may change the result
-  const enableCache =
-    !isSimpleType(context.program, type) && !isEmptyModel(type);
+  const enableCache = !isSimpleType(context.program, type);
   const effectiveModel =
     !options.disableEffectiveModel &&
     (type.kind === "Model" || type.kind === "Union")
@@ -298,8 +308,14 @@ function getType(
     newValue = emitType(context, type);
   }
 
+  if (type.kind === "ModelProperty" || type.kind === "Scalar") {
+    newValue = applyEncoding(context.program, type, newValue);
+  }
+
   if (enableCache) {
-    typesMap.set(effectiveModel, newValue);
+    if (!options.disableEffectiveModel) {
+      typesMap.set(effectiveModel, newValue);
+    }
     if (type.kind === "Union") {
       for (const t of type.variants.values()) {
         if (t.type.kind === "Model") {
@@ -335,6 +351,7 @@ type ParamBase = {
   addedOn: string | undefined;
   clientName: string;
   inOverload: boolean;
+  format?: string;
 };
 function emitParamBase(
   program: Program,
@@ -344,12 +361,15 @@ function emitParamBase(
   let name: string;
   let description: string = "";
   let addedOn: string | undefined;
+  let format: string | undefined;
 
   if (parameter.kind === "ModelProperty") {
+    const newParameter = applyEncoding(program, parameter, parameter);
     optional = parameter.optional;
     name = parameter.name;
     description = getDocStr(program, parameter);
     addedOn = getAddedOnVersion(program, parameter);
+    format = newParameter.format;
   } else {
     optional = false;
     name = "body";
@@ -360,7 +380,8 @@ function emitParamBase(
     description,
     addedOn,
     clientName: applyCasing(name, { casing: CASING }),
-    inOverload: false
+    inOverload: false,
+    format
   };
 }
 
@@ -429,10 +450,6 @@ function emitBodyParameter(
     disableEffectiveModel: true
   });
 
-  if (type.type === "model" && type.name === "") {
-    type.name = capitalize(httpOperation.operation.name) + "Request";
-  }
-
   return {
     contentTypes,
     type,
@@ -466,10 +483,10 @@ function emitParameter(
   const paramMap: any = {
     restApiName: parameter.name,
     location: parameter.type,
-    type: type,
+    type: base.format ? { ...type, format: base.format } : type,
     implementation: implementation,
     skipUrlEncoding: parameter.type === "endpointPath",
-    format: (parameter as any).format
+    format: (parameter as any).format ?? base.format
   };
 
   if (paramMap.type.type === "constant") {
@@ -606,7 +623,10 @@ function emitOperation(
   operationGroupName: string,
   rlcModels: RLCModel
 ): HrlcOperation {
-  const lro = isLro(context.program, operation);
+  const lro = isLongRunningOperation(
+    context.program,
+    ignoreDiagnostics(getHttpOperation(context.program, operation))
+  );
   const paging = getPagedResult(context.program, operation);
   if (lro && paging) {
     return emitLroPagingOperation(
@@ -777,7 +797,17 @@ function emitBasicOperation(
     bodyParameter = undefined;
   } else {
     bodyParameter = emitBodyParameter(context, httpOperation);
-    if (
+    if (bodyParameter.type.type === "model" && bodyParameter.type.name === "") {
+      if (bodyParameter.type.properties.length > 0) {
+        for (const param of bodyParameter.type.properties) {
+          // const emittedParam = emitParameter(context, param.type, "Method");
+          param.implementation = "Method";
+          param.location = param.location ?? "body";
+          parameters.push(param);
+        }
+      }
+      bodyParameter = undefined;
+    } else if (
       bodyParameter.type.type === "model" &&
       bodyParameter.type.base === "json"
     ) {
@@ -793,6 +823,21 @@ function emitBasicOperation(
     }
   }
   const name = applyCasing(operation.name, { casing: CASING });
+
+  /** handle name collision between operation name and parameter signature */
+  if (bodyParameter) {
+    bodyParameter.clientName =
+      bodyParameter.clientName === name
+        ? bodyParameter.clientName + "Parameter"
+        : bodyParameter.clientName;
+  }
+  parameters
+    .filter((param) => {
+      return param.clientName === name && !param.isReadOnly && param.required;
+    })
+    .forEach((param) => {
+      param.clientName = param.clientName + "Parameter";
+    });
   return {
     name: name,
     description: getDocStr(context.program, operation),
@@ -828,6 +873,7 @@ function emitProperty(
   context: SdkContext,
   property: ModelProperty
 ): Record<string, any> {
+  const newProperty = applyEncoding(context.program, property, property);
   let clientDefaultValue = undefined;
   const propertyDefaultKind = property.default?.kind;
   if (
@@ -851,16 +897,18 @@ function emitProperty(
   if (property.model) {
     getType(context, property.model);
   }
+  const type = getType(context, property.type);
   return {
     clientName: applyCasing(clientName, { casing: CASING }),
     restApiName: jsonName,
-    type: getType(context, property.type),
+    type: newProperty.format ? { ...type, format: newProperty.format } : type,
     optional: property.optional,
     description: getDocStr(context.program, property),
     addedOn: getAddedOnVersion(context.program, property),
     readonly:
       isReadOnly(context.program, property) || isKey(context.program, property),
-    clientDefaultValue: clientDefaultValue
+    clientDefaultValue: clientDefaultValue,
+    format: newProperty.format
   };
 }
 
@@ -1004,6 +1052,14 @@ function emitCredential(auth: HttpAuth): Record<string, any> {
         key: auth.name
       }
     };
+  } else if (auth.type === "http") {
+    credential_type = {
+      type: "Key",
+      policy: {
+        type: "AzureKeyCredentialPolicy",
+        key: "Authorization"
+      }
+    };
   }
   return credential_type;
 }
@@ -1024,9 +1080,10 @@ function emitStdScalar(
   program: Program,
   scalar: Scalar & { name: IntrinsicScalarName }
 ): Record<string, any> {
+  const newScalar = applyEncoding(program, scalar, scalar);
   switch (scalar.name) {
     case "bytes":
-      return { type: "byte-array", format: getEncode(program, scalar) };
+      return { type: "byte-array", format: newScalar.format };
     case "int8":
     case "int16":
     case "int32":
@@ -1048,17 +1105,58 @@ function emitStdScalar(
     case "boolean":
       return { type: "boolean" };
     case "plainDate":
-      return { type: "date" };
+      return { type: "datetime", format: newScalar.format ?? "date" };
     case "utcDateTime":
-      return { type: "datetime", format: "date-time" };
+      return { type: "datetime", format: newScalar.format };
     case "plainTime":
-      return { type: "time" };
+      return { type: "datetime", format: newScalar.format ?? "time" };
+    case "offsetDateTime":
+      return { type: "string" };
     case "duration":
-      return { type: "duration" };
+      return { type: "duration", format: newScalar.format };
     case "numeric":
       return {}; // Waiting on design for more precise type https://github.com/microsoft/cadl/issues/1260
     default:
       return {};
+  }
+}
+
+function applyEncoding(
+  program: Program,
+  typespecType: Scalar | ModelProperty,
+  target: any = {}
+) {
+  const encodeData = getEncode(program, typespecType);
+  const formatData = getFormat(program, typespecType);
+  formatData;
+  if (encodeData) {
+    const newTarget = { ...target };
+    const newType = emitScalar(program, encodeData.type);
+    // newTarget["type"] = newType["type"];
+    // If the target already has a format it takes priority. (e.g. int32)
+    newTarget["format"] = mergeFormatAndEncoding(
+      newTarget.format,
+      encodeData.encoding,
+      newType["format"]
+    );
+    return newTarget;
+  }
+  return target;
+}
+
+function mergeFormatAndEncoding(
+  format: string | undefined,
+  encoding: string,
+  encodeAsFormat: string | undefined
+): string {
+  switch (format) {
+    case undefined:
+      return encodeAsFormat ?? encoding;
+    case "date-time":
+      return encoding;
+    case "duration":
+    default:
+      return encodeAsFormat ?? encoding;
   }
 }
 
@@ -1067,7 +1165,7 @@ function applyIntrinsicDecorators(
   type: Scalar | ModelProperty,
   result: any
 ): Record<string, any> {
-  const newResult = { ...result };
+  let newResult = { ...result };
   const docStr = getDoc(program, type);
   const isString = isStringType(program, getPropertyType(type));
   const isNumeric = isNumericType(program, getPropertyType(type));
@@ -1076,10 +1174,7 @@ function applyIntrinsicDecorators(
     newResult.description = docStr;
   }
 
-  const formatStr = getFormat(program, type);
-  if (isString && !result.format && formatStr) {
-    newResult.format = formatStr;
-  }
+  newResult = applyEncoding(program, type, newResult);
 
   const pattern = getPattern(program, type);
   if (isString && !result.pattern && pattern) {
@@ -1120,7 +1215,8 @@ function applyIntrinsicDecorators(
 
 function emitScalar(program: Program, scalar: Scalar): Record<string, any> {
   let result: Record<string, any> = {};
-  if (program.checker.isStdType(scalar)) {
+  const isStd = program.checker.isStdType(scalar);
+  if (isStd) {
     result = emitStdScalar(program, scalar);
   } else if (scalar.baseScalar) {
     result = emitScalar(program, scalar.baseScalar);
@@ -1164,12 +1260,9 @@ function mapTypeSpecType(context: SdkContext, type: Type): any {
   }
 }
 
-function capitalize(name: string): string {
-  return name[0]!.toUpperCase() + name.slice(1);
-}
-
 function emitUnion(context: SdkContext, type: Union): Record<string, any> {
   const sdkType = getSdkUnion(context, type);
+  const nonNullOptions = getNonNullOptions(type);
   if (sdkType === undefined) {
     throw Error("Should not have an empty union");
   }
@@ -1181,29 +1274,65 @@ function emitUnion(context: SdkContext, type: Union): Record<string, any> {
       description: `Type of ${unionName}`,
       internal: true,
       type: "combined",
-      types: sdkType.values.map((x) => getType(context, x.__raw)),
+      types: sdkType.values.map((x) => getType(context, x.__raw!)),
       xmlMetadata: {}
     };
   } else if (sdkType.kind === "enum") {
     return {
       name: sdkType.name,
       nullable: sdkType.nullable,
-      description: sdkType.doc || `Type of ${sdkType.name}`,
+      description: sdkType.description || `Type of ${sdkType.name}`,
       internal: true,
       type: sdkType.kind,
-      valueType: emitSimpleType(context, sdkType.valueType as SdkSimpleType),
+      valueType: emitSimpleType(context, sdkType.valueType as SdkBuiltInType),
       values: sdkType.values.map((x) => emitEnumMember(x)),
-      isFixed: sdkType.isFixed === false ? false : true,
+      isFixed: sdkType.isFixed,
+      xmlMetadata: {}
+    };
+  } else if (
+    isStringOrNumberKind(context.program, sdkType.kind) &&
+    isStringOrNumberKind(context.program, nonNullOptions[0]?.kind)
+  ) {
+    return {
+      name: undefined, // string/number union will be mapped as fixed enum without name
+      nullable: sdkType.nullable,
+      internal: true,
+      type: "enum",
+      valueType: emitSimpleType(context, sdkType as SdkBuiltInType),
+      values: nonNullOptions
+        .map((x) => getClientType(context, x))
+        .map((x) => emitEnumMember(x)),
+      isFixed: true,
       xmlMetadata: {}
     };
   } else {
-    return { nullable: sdkType.nullable, ...emitType(context, sdkType.__raw) };
+    return { nullable: sdkType.nullable, ...emitType(context, sdkType.__raw!) };
   }
+}
+
+function isStringOrNumberKind(program: Program, kind?: string): boolean {
+  if (!kind) {
+    return false;
+  }
+  kind = kind.toLowerCase();
+  try {
+    const type = emitStdScalar(program, { name: kind } as any);
+    kind = type["type"] ?? kind;
+  } catch (e: any) {
+    // ignore
+  }
+  return ["string", "number", "integer", "float"].includes(kind!);
+}
+
+function getNonNullOptions(type: Union) {
+  return [...type.variants.values()]
+    .map((x) => x.type)
+    .filter((t) => !isNullType(t));
 }
 
 function emitEnumMember(type: any): Record<string, any> {
   return {
-    name: enumName(type.name),
+    name: type.name ? enumName(type.name) : undefined,
     value: type.value,
     description: type.doc
   };
@@ -1211,37 +1340,22 @@ function emitEnumMember(type: any): Record<string, any> {
 
 function emitSimpleType(
   context: SdkContext,
-  type: Scalar | IntrinsicType | SdkSimpleType
+  type: Scalar | IntrinsicType | SdkBuiltInType
 ): Record<string, any> {
-  let sdkType: SdkSimpleType;
+  let sdkType: SdkBuiltInType;
   if (type.kind === "Scalar" || type.kind === "Intrinsic") {
-    sdkType = getSdkSimpleType(context, type);
+    sdkType = getSdkBuiltInType(context, type);
   } else {
     sdkType = type;
   }
 
-  const extraInformation: Record<string, any> = {};
-  if (sdkType.kind === "string") {
-    extraInformation["pattern"] = sdkType.pattern;
-    extraInformation["minLength"] = sdkType.minLength;
-    extraInformation["maxLength"] = sdkType.maxLength;
-  } else if (
-    sdkType.kind === "int32" ||
-    sdkType.kind === "int64" ||
-    sdkType.kind === "float32" ||
-    sdkType.kind === "float64"
-  ) {
-    extraInformation["minValue"] = sdkType.minValue;
-    extraInformation["maxValue"] = sdkType.maxValue;
-  }
   return {
     nullable: sdkType.nullable,
     type: "number", // TODO: switch to kind
-    doc: sdkType.doc,
-    apiVersions: sdkType.apiVersions,
-    sdkDefaultValue: sdkType.sdkDefaultValue,
-    format: sdkType.format,
-    ...extraInformation
+    doc: "",
+    apiVersions: [],
+    sdkDefaultValue: undefined,
+    format: undefined
   };
 }
 
@@ -1575,9 +1689,11 @@ export function emitCodeModel(
     project
   };
 
+  typesMap.clear();
+  simpleTypesMap.clear();
   const allModels = getAllModels(dpgContext);
   for (const model of allModels) {
-    getType(dpgContext, model);
+    getType(dpgContext, model.__raw!);
   }
 
   for (const namespace of getNamespaces(dpgContext)) {