+++ packages/typespec-ts/src/modular/buildOperations.ts
@@ -23,6 +23,7 @@ export function buildOperationFiles(
   client: Client,
   needUnexpectedHelper: boolean = true
 ) {
+  const operationFiles = [];
   for (const operationGroup of client.operationGroups) {
     const importSet: Map<string, Set<string>> = new Map<string, Set<string>>();
     const fileName = operationGroup.className
@@ -114,7 +115,9 @@ export function buildOperationFiles(
     operationGroupFile.fixMissingImports();
     // have to fixUnusedIdentifiers after everything get generated.
     operationGroupFile.fixUnusedIdentifiers();
+    operationFiles.push(operationGroupFile);
   }
+  return operationFiles;
 }
 
 export function importModels(
@@ -156,10 +159,7 @@ export function buildOperationOptions(
   const optionalParameters = operation.parameters
     .filter((p) => p.implementation === "Method")
     .filter((p) => p.optional || p.clientDefaultValue);
-  const optionalBodyParams = (
-    operation.bodyParameter?.type.properties ?? []
-  ).filter((p) => p.optional);
-  const options = [...optionalBodyParams, ...optionalParameters];
+  const options = [...optionalParameters];
 
   const name = getOperationOptionsName(operation);
 
@@ -171,7 +171,7 @@ export function buildOperationOptions(
       return {
         docs: getDocsFromDescription(p.description),
         hasQuestionToken: true,
-        ...buildType(p.clientName, p.type)
+        ...buildType(p.clientName, p.type, p.format)
       };
     })
   });