+++ packages/typespec-ts/src/modular/helpers/clientHelpers.ts
@@ -1,4 +1,8 @@
-import { OptionalKind, ParameterDeclarationStructure } from "ts-morph";
+import {
+  OptionalKind,
+  ParameterDeclarationStructure,
+  SourceFile
+} from "ts-morph";
 import { Client } from "../modularCodeModel.js";
 import { getType } from "./typeHelpers.js";
 import { getClientName } from "./namingHelpers.js";
@@ -24,11 +28,18 @@ export function getClientParameters(
       .map<OptionalKind<ParameterDeclarationStructure>>((p) => {
         return {
           name: p.clientName,
-          type: getType(p.type).name
+          type: getType(p.type, p.format).name
         };
       }),
     optionsParam
   ];
 
   return params;
+}
+
+export function importCredential(clientSourceFile: SourceFile): void {
+  clientSourceFile.addImportDeclaration({
+    moduleSpecifier: "@azure/core-auth",
+    namedImports: ["TokenCredential", "KeyCredential"]
+  });
 }